Added Rename
Added rename capability to the VFS, this required implementing the
remove capability in the memfs

+++ temp.go
@@ -3,23 +3,26 @@ package vfs
 import (
 	"io/ioutil"
 	"os"
-	"path/filepath"
 )
 
 type tempfs struct {
-	*OsFs
+	FileSystem
 	tempdir string
 }
 
 func NewTempFs() FileSystem {
 	tempdir, _ := ioutil.TempDir("", "osfs_test")
 	return &tempfs{
-		OsFs:    &OsFs{filepath.Clean(tempdir)},
-		tempdir: tempdir,
+		FileSystem: NewOsFs(tempdir),
+		tempdir:    tempdir,
 	}
 }
 
 func (tfs *tempfs) Close() error {
-	tfs.OsFs = nil
-	return os.RemoveAll(tfs.tempdir)
+	err := tfs.FileSystem.Close()
+	if err == nil {
+		tfs.FileSystem = nil
+		err = os.RemoveAll(tfs.tempdir)
+	}
+	return err
 }
